Style change

+++ include/tradfri.h
@@ -9,6 +9,6 @@ int TradfriGetAllLamps(char* response);
 int TradfriGetLamp(char* lampId, char *response);
 int TradfriTurnOnLamp(char* lampId, char* response);
 int TradfriTurnOffLamp(char* lampId, char* response);
-int TradfriDimLamp(char* lampId, int dim, char* response);
+int TradfriDimLamp(char* lampId, uint8_t dim, char* response);
 int TradfriSetLampColor(char* lampId, uint64_t color_hex, char *response);
 #endif
